RegisterAxis refactor
Refactored the way axis inputs are registered so they store the strings instead of forming the strings each frame. Created a new GamepadInputRegistrar to separate the tie that bindings had to use a string so keyboard input now works again (it has its own dictionary using KeyCode as the value while PS4, Xbox and Samsung registrars inherit from the new gamepad registrar which uses a string).